Create README.md
Our data analysis will focus on iPhone phones sold on https://www.amazon.fr website to pinpoint key patterns or specific behaviors. We will use Web Scraping which makes reference to the retrieval of data from specific pages of a web site. This data is picked up and then exported in a more useful format for the user, in our case, in an excel sheet.